requirements.txt
certifi==2020.6.20
chardet==3.0.4
dadjokes==1.3.2
falcon==2.0.0
falcon-auth==1.1.0
gunicorn==20.0.4
idna==2.10
PyJWT==1.7.1
python-dotenv==0.14.0
requests==2.24.0
urllib3==1.25.9
